requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
python-decouple==3.8
python-dotenv==1.0.1
Werkzeug==3.1.2
WTForms==3.2.1
